refactor(chatbot): restructure Bots directory with handlers and interfaces for extensibility
